feat/password-hashing: More function migration so the compiler stops shouting me
